requirements.txt
httpx[async]>=0.27
schedule>=1.2
jinja2>=3.1
python-dotenv>=1.0
